refactor: moved files from stock_entry_handler to services
